Update lemmatize and added features
I have added a new feature to consider flex verbs that have duplicity against others. For instance, if a flex verb exists in more than one infinitive verb list, it is considered duplicated and therefore will not be lemmatized. If you want to be notified when this occurs, you can set the parameter 'silence' to 'False' in the 'lemmatize' function.

Additionally, another function called 'checkFlexVerbDuplicity' has been added, allowing you to search for duplicated verbs found in the database I collected.