Update features for flex verbs ambiguity
As this tool is "rule-based", when a flex verb has more than just one infinitive verb associated with it,  the lemmatization isn't precise, the tool doesn't know what's the correct association between, for example, "podem": is it associated with "poder" or "podar"? Both are correct if we think in a rule-based limitation. So I developt two other functions that are willing to decrease the failure, which are "checkFlexVerbDuplicity()" and "modifyDicAmbiguousVerbs()".